fixed (double)s

+++ KEARumfangOverfladeKugle/Form1.cs
@@ -29,8 +29,10 @@ namespace KEARumfangOverfladeKugle
                 labelError.Text = "Only use numbers."; // error
             }
 
-            volume = 4 / 3 * Math.PI * Math.Pow(r, 3); // supposed formula for volume of a ball
-            area = 4 * Math.PI * Math.Pow(r, 2); // supposed formula for area of a ball
+            // use '(double)' in front of digits to make them a double instead of int.
+            // OR write it out with comma: (double)4 or 4.0
+            volume = (double)4 / 3.0 * Math.PI * Math.Pow(r, 3); // supposed formula for volume of a ball
+            area = 4.0 * Math.PI * Math.Pow(r, 2); // supposed formula for area of a ball
 
             textBoxVolume.Text = Convert.ToString(Math.Round(volume, 2)); // round down to 2 decimals and converts to string and inputs to textbox
             textBoxArea.Text = Convert.ToString(Math.Round(area, 2)); // ------ || -----
